+++ src/pack/pack_utils.hpp
@@ -16,6 +16,7 @@
 #include <string>
 #include <tuple>
 #include <type_traits>
+#include <unordered_map>
 #include <utility>
 #include <vector>
 
@@ -53,6 +54,9 @@ struct GetDataType<T> {
 } // namespace
 
 namespace parthenon {
+
+// Map for going from variable names to pack variable indices
+using PackIdxMap = std::unordered_map<std::string, std::size_t>;
 
 // Namespace in which to put variable name types that are used for indexing into
 // SparsePack<[type list of variable name types]> on device